[move] Extract bytecode-verifier test helper
`dummy_procedure_module` is only used by bytecode-verifier-tests. Place
it in a test support module close to its only usages, in order to
discourage it from being used elsewhere.

Closes: #8735
