+++ language/bytecode-verifier/bytecode-verifier-tests/src/unit_tests/code_unit_tests.rs
@@ -1,13 +1,14 @@
 // Copyright (c) The Diem Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::support::dummy_procedure_module;
 use bytecode_verifier::CodeUnitVerifier;
-use move_binary_format::file_format::{self, Bytecode};
+use move_binary_format::file_format::Bytecode;
 use move_core_types::vm_status::StatusCode;
 
 #[test]
 fn invalid_fallthrough_br_true() {
-    let module = file_format::dummy_procedure_module(vec![Bytecode::LdFalse, Bytecode::BrTrue(1)]);
+    let module = dummy_procedure_module(vec![Bytecode::LdFalse, Bytecode::BrTrue(1)]);
     let result = CodeUnitVerifier::verify_module(&module);
     assert_eq!(
         result.unwrap_err().major_status(),
@@ -17,7 +18,7 @@ fn invalid_fallthrough_br_true() {
 
 #[test]
 fn invalid_fallthrough_br_false() {
-    let module = file_format::dummy_procedure_module(vec![Bytecode::LdTrue, Bytecode::BrFalse(1)]);
+    let module = dummy_procedure_module(vec![Bytecode::LdTrue, Bytecode::BrFalse(1)]);
     let result = CodeUnitVerifier::verify_module(&module);
     assert_eq!(
         result.unwrap_err().major_status(),
@@ -28,7 +29,7 @@ fn invalid_fallthrough_br_false() {
 // all non-branch instructions should trigger invalid fallthrough; just check one of them
 #[test]
 fn invalid_fallthrough_non_branch() {
-    let module = file_format::dummy_procedure_module(vec![Bytecode::LdTrue, Bytecode::Pop]);
+    let module = dummy_procedure_module(vec![Bytecode::LdTrue, Bytecode::Pop]);
     let result = CodeUnitVerifier::verify_module(&module);
     assert_eq!(
         result.unwrap_err().major_status(),
@@ -38,21 +39,21 @@ fn invalid_fallthrough_non_branch() {
 
 #[test]
 fn valid_fallthrough_branch() {
-    let module = file_format::dummy_procedure_module(vec![Bytecode::Branch(0)]);
+    let module = dummy_procedure_module(vec![Bytecode::Branch(0)]);
     let result = CodeUnitVerifier::verify_module(&module);
     assert!(result.is_ok());
 }
 
 #[test]
 fn valid_fallthrough_ret() {
-    let module = file_format::dummy_procedure_module(vec![Bytecode::Ret]);
+    let module = dummy_procedure_module(vec![Bytecode::Ret]);
     let result = CodeUnitVerifier::verify_module(&module);
     assert!(result.is_ok());
 }
 
 #[test]
 fn valid_fallthrough_abort() {
-    let module = file_format::dummy_procedure_module(vec![Bytecode::LdU64(7), Bytecode::Abort]);
+    let module = dummy_procedure_module(vec![Bytecode::LdU64(7), Bytecode::Abort]);
     let result = CodeUnitVerifier::verify_module(&module);
     assert!(result.is_ok());
 }
