+++ language/bytecode-verifier/bytecode-verifier-tests/src/unit_tests/control_flow_tests.rs
@@ -1,11 +1,12 @@
 // Copyright (c) The Diem Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::support::dummy_procedure_module;
 use bytecode_verifier::control_flow;
 use move_binary_format::{
     access::ModuleAccess,
     errors::PartialVMResult,
-    file_format::{self, Bytecode, CompiledModule, FunctionDefinitionIndex, TableIndex},
+    file_format::{Bytecode, CompiledModule, FunctionDefinitionIndex, TableIndex},
 };
 use move_core_types::vm_status::StatusCode;
 
@@ -33,7 +34,7 @@ fn verify_module(module: &CompiledModule) -> PartialVMResult<()> {
 
 #[test]
 fn invalid_fallthrough_br_true() {
-    let module = file_format::dummy_procedure_module(vec![Bytecode::LdFalse, Bytecode::BrTrue(1)]);
+    let module = dummy_procedure_module(vec![Bytecode::LdFalse, Bytecode::BrTrue(1)]);
     let result = verify_module(&module);
     assert_eq!(
         result.unwrap_err().major_status(),
@@ -43,7 +44,7 @@ fn invalid_fallthrough_br_true() {
 
 #[test]
 fn invalid_fallthrough_br_false() {
-    let module = file_format::dummy_procedure_module(vec![Bytecode::LdTrue, Bytecode::BrFalse(1)]);
+    let module = dummy_procedure_module(vec![Bytecode::LdTrue, Bytecode::BrFalse(1)]);
     let result = verify_module(&module);
     assert_eq!(
         result.unwrap_err().major_status(),
@@ -54,7 +55,7 @@ fn invalid_fallthrough_br_false() {
 // all non-branch instructions should trigger invalid fallthrough; just check one of them
 #[test]
 fn invalid_fallthrough_non_branch() {
-    let module = file_format::dummy_procedure_module(vec![Bytecode::LdTrue, Bytecode::Pop]);
+    let module = dummy_procedure_module(vec![Bytecode::LdTrue, Bytecode::Pop]);
     let result = verify_module(&module);
     assert_eq!(
         result.unwrap_err().major_status(),
@@ -64,21 +65,21 @@ fn invalid_fallthrough_non_branch() {
 
 #[test]
 fn valid_fallthrough_branch() {
-    let module = file_format::dummy_procedure_module(vec![Bytecode::Branch(0)]);
+    let module = dummy_procedure_module(vec![Bytecode::Branch(0)]);
     let result = verify_module(&module);
     assert!(result.is_ok());
 }
 
 #[test]
 fn valid_fallthrough_ret() {
-    let module = file_format::dummy_procedure_module(vec![Bytecode::Ret]);
+    let module = dummy_procedure_module(vec![Bytecode::Ret]);
     let result = verify_module(&module);
     assert!(result.is_ok());
 }
 
 #[test]
 fn valid_fallthrough_abort() {
-    let module = file_format::dummy_procedure_module(vec![Bytecode::LdU64(7), Bytecode::Abort]);
+    let module = dummy_procedure_module(vec![Bytecode::LdU64(7), Bytecode::Abort]);
     let result = verify_module(&module);
     assert!(result.is_ok());
 }
