+++ language/move-binary-format/src/file_format.rs
@@ -1972,31 +1972,6 @@ pub fn basic_test_module() -> CompiledModule {
     m
 }
 
-/// Create a dummy module to wrap the bytecode program in local@code
-pub fn dummy_procedure_module(code: Vec<Bytecode>) -> CompiledModule {
-    let mut module = empty_module();
-    let code_unit = CodeUnit {
-        code,
-        ..Default::default()
-    };
-    let fun_def = FunctionDefinition {
-        code: Some(code_unit),
-        ..Default::default()
-    };
-
-    let fun_handle = FunctionHandle {
-        module: ModuleHandleIndex(0),
-        name: IdentifierIndex(0),
-        parameters: SignatureIndex(0),
-        return_: SignatureIndex(0),
-        type_parameters: vec![],
-    };
-
-    module.function_handles.push(fun_handle);
-    module.function_defs.push(fun_def);
-    module.freeze().unwrap()
-}
-
 /// Return a simple script that contains only a return in the main()
 pub fn empty_script() -> CompiledScript {
     CompiledScript {
